@@ -4,21 +4,21 @@ version: 1.0.1
 homepage: https://kodefabrique.com
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ^3.0.0
   flutter: ">=2.10.5"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  webview_flutter: ^4.0.2
+  webview_flutter: ^4.7.0
   pkce: ^1.1.0+1
-  dio: ^4.0.4
+  dio: ^5.4.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^3.0.1
 
 flutter:
 
